Version 0.1.8
Imported WPF template from an other project into TravelWindow.xaml. The left grid contains textboxes, labels and a button with logic that could be used as "add travels" window. The right grid could be used in the travelwindow as is, to display the various trips/vacations the current logged on user already added to its list. The right grid could also be used to click on any item/trip and show details, in which the button "Show details" opens up TravelDetailsWindow, and the other button "remove" does... well, deletes.