Version 0.2.4
Working around the current problem which is that I cannot log in with default users such as Gandalf and Admin.

In the next version I will hardcode two trips for Gandalf, I hope to be able to log in as Gandalf, in order to further see the code in action.

I have created a TravelManager, introduced properties to it and also introduced properties to the Travel class. Also created new classes Trip and Vacation that inherit from Travel.