Version 0.4.6
Fixed issue where trip type combobox and All inclusive checkbox would not appear when choosing the specific type of travel.

Validate username is implemented. You cannot register a user with an already taken username

Cancel button added to userdetailswindow that will revert changes and go back to travelswindow.

Cleaner code when trying to sign-in. Removed the foreach loop from register window as this already occurs in the UserManager.